Tidy up; rearranged HashComparer.
